A little bit refactoring
