+++ for Arduino/photopizza/preset.cpp
@@ -1,0 +1,38 @@
+/**
+ * File: preset.cpp
+ * Created on: 14 марта 2015 г.
+ * Description:
+ * 
+ *
+ * Author: Roman Savrulin <romeo.deepmind@gmail.com>
+ * Project Author:
+ * Project Site: PhotoPizza.org
+ *
+ * Copyright: 2015 Roman Savrulin
+ * Copying permission statement:
+ * 
+ *  This file is part of PhotoPizza DIY.
+ *
+ *  PhotoPizza DIY is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ *
+ */
+
+#include "preset.h"
+
+using namespace PhotoPizza;
+
+/* static */ LimitedParam preset::_default;
+/* static */ paramRun preset::_run;
+
