+++ for Arduino/photopizza/presetManager.h
@@ -31,128 +31,9 @@
 #define PRESET_MANAGER_H
 
 #include "defines.h"
-#include "eepromAnything.h"
-
-#include "param.h"
-#include "paramAcc.h"
-#include "paramSteps.h"
-#include "paramSpeed.h"
-#include "paramDir.h"
-#include "paramRun.h"
-#include "paramIterCount.h"
-#include "paramPause.h"
+#include "preset.h"
 
 namespace PhotoPizza {
-
-struct presetStorageData {
-  long _steps;  // rotaion
-  long _accel;    // acceleration
-  long _pause;
-  int   _iter;
-  short _speed; // speed
-};
-
-struct presetStorage {
-  byte flag;
-  int  version;
-
-  presetStorageData data;
-};
-
-class preset {
-public:
-  paramSpeed     _speed; // speed
-  paramSteps     _steps; // rotaion
-  paramAcc       _acc; // acceleration
-  paramDir       _dir; // -1 - clockwise , 1 - counterclockwise
-  paramIterCount _iter;
-  paramPause     _pause;
-
-  static paramRun _run;
-
-  typedef enum {
-    FIRST_PARAM = 0,
-    RUN = FIRST_PARAM,  // command-parameter
-    STEPS,
-    SPEED,
-    ITER_COUNT,
-    PAUSE,
-    ACC,
-    DIR,
-    PARAM_COUNT,
-
-    SAVED_PARAM,       //parameter where we save value before editing
-  } paramType;
-
-  IParam& operator[] (const int nIndex){
-    switch (nIndex) {
-    case RUN:
-      return _run;
-    case SPEED:
-      return _speed;
-    case ITER_COUNT:
-      return _iter;
-    case PAUSE:
-        return _pause;
-    case STEPS:
-      return _steps;
-    case ACC:
-      return _acc;
-    case DIR:
-      return _dir;
-    default:
-      return _default;
-    }
-  }
-
-  bool operator!=(presetStorageData & val) {
-    if(_speed != val._speed)
-       return true;
-    if(_acc != val._accel)
-       return true;
-    if(_steps * _dir != val._steps)
-       return true;
-    if(_pause != val._pause)
-      return true;
-    if(_iter != val._iter)
-        return true;
-
-    return false;
-  }
-
-  operator presetStorageData() {
-    //Serial.println("op = cast to psd");
-    presetStorageData tmp;
-    tmp._accel = _acc;
-    tmp._speed = _speed;
-    tmp._steps = _steps * _dir;
-    tmp._iter = _iter;
-    tmp._pause = _pause;
-    return tmp;
-  }
-
-  preset& operator=(presetStorageData val) {
-//    Serial.println(F("op = psd"));
-//    Serial.println((String)F("Sp: ") + val._speed);
-//    Serial.println((String)F("acc: ") + val._acc);
-//    Serial.println((String)F("steps: ") + val._steps);
-    _speed = val._speed;
-    _acc = val._accel;
-    _iter = val._iter;
-    _pause = val._pause;
-
-    if(val._steps < 0)
-      _dir.setByVal(CW);
-    else
-      _dir.setByVal(CCW);
-
-    _steps = abs(val._steps);
-    return *this;
-  }
-
-private:
-  static LimitedParam _default;
-};
 
 class presetManager {
 public:
@@ -211,7 +92,6 @@ public:
 private:
   bool loadPreset(unsigned short num, bool set = false);
   bool savePreset(unsigned short num);
-  static presetManager *_presetMgr;
   presetManager();
   int _curPreset; // current preset
   preset _preset;
